requirements.txt
mysql-connector-python
prettytable
tabulate
uuid
getpass4